Extract setting data to separate module
